use decimal for balance & price

+++ balance-mng/balance-mng-facade/src/main/java/io/sofastack/balance/manage/facade/BalanceMngFacade.java
@@ -19,7 +19,7 @@ public interface BalanceMngFacade {
      *
      * @param userName 用户名
      */
-    int queryBalance(String userName);
+    Double queryBalance(String userName);
 
     /**
      * 减少用户余额
